build(client-c): migrate example build system to CMake with strict warnings
- Replace direct compiler scripts with a CMake build configuration.
- Introduce an interface target for extensible dependency and include path management.
- Configure compiler flags with strict warnings (`/W4` on MSVC, `-Wall -Wextra -Wpedantic` on GCC/Clang).
- Add `build_vs2022.bat` and update `build.sh` to execute CMake build steps with clean builds.
- Fix implicit conversion warnings in `example_cpp.cpp` using explicit `static_cast`.

+++ rt-client-c/examples/example_cpp.cpp
@@ -57,8 +57,8 @@ int main(int argc, char* argv[]) {
         // clang-format off
         std::string arg = argv[i];
         if (arg == "--host" && i + 1 < argc) host = argv[++i];
-        else if (arg == "--port" && i + 1 < argc) port = std::stoi(argv[++i]);
-        else if (arg == "--client_id" && i + 1 < argc) client_id = std::stoi(argv[++i]);
+        else if (arg == "--port" && i + 1 < argc) port = static_cast<uint16_t>(std::stoi(argv[++i]));
+        else if (arg == "--client_id" && i + 1 < argc) client_id = static_cast<uint16_t>(std::stoi(argv[++i]));
         else if (arg == "--run_cycle_sec" && i + 1 < argc) run_cycle_sec = std::stod(argv[++i]);
         else if (arg == "--startup_wait_sec" && i + 1 < argc) startup_wait_sec = std::stod(argv[++i]);
         else if (arg == "--proc_time_sec" && i + 1 < argc) proc_time_sec = std::stod(argv[++i]);
